chore(test): add micro-benchmark performance test for Observable.zip
- move Observable.zip test from operator test location
- add Observable.prototype.zip test to operator test location
- remove zip test from current thread tests
