+++ perf/micro/immediate-scheduler/operators/zip.js
@@ -3,14 +3,14 @@ var RxNew = require("../../../../index");
 
 module.exports = function (suite) {
     
-    var oldZipWithImmediateScheduler = RxOld.Observable.zip(RxOld.Observable.range(0, 25), RxOld.Observable.range(0, 25), add);
-    var newZipWithImmediateScheduler = RxNew.Observable.zip(RxNew.Observable.range(0, 25), RxNew.Observable.range(0, 25), add);
+    var oldZipWithImmediateScheduler = RxOld.Observable.range(0, 25).zip(RxOld.Observable.range(0, 25), add);
+    var newZipWithImmediateScheduler = RxNew.Observable.range(0, 25).zip(RxNew.Observable.range(0, 25), add);
 
     return suite
-        .add('old zip with immediate scheduler', function () {
+        .add('old zip() with immediate scheduler', function () {
             oldZipWithImmediateScheduler.subscribe(_next, _error, _complete);
         })
-        .add('new zip with immediate scheduler', function () {
+        .add('new zip() with immediate scheduler', function () {
             newZipWithImmediateScheduler.subscribe(_next, _error, _complete);
         });
